frontend-utils: Move WinitAsyncExecutor from Desktop
