+++ desktop/Cargo.toml
@@ -27,7 +27,6 @@ ruffle_frontend_utils = { path = "../frontend-utils" }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing-appender = "0.2.3"
-slotmap = { workspace = true }
 winit = "0.29.15"
 webbrowser = "0.8.14"
 url = { workspace = true }
