+++ frontend-utils/Cargo.toml
@@ -18,6 +18,8 @@ thiserror = { workspace = true }
 zip = "0.6.6"
 urlencoding = "2.1.3"
 ruffle_core = { path = "../core", default-features = false }
+async-channel = { workspace = true }
+slotmap = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
